dating backend/requirements.txt
Django
gunicorn
gevent
psycopg2-binary
dj-database-url
python-dotenv
whitenoise
django-cors-headers
djangorestframework
python-dateutil
google-auth
razorpay
cohere
pandas
matplotlib
agora-token-builder
Pillow
cloudinary
django-cloudinary-storage
